Build the run-time into libverilator_rt*.a ahead of time
