+++ .gitignore
@@ -40,3 +40,4 @@ verilator-config-version.cmake
 **/_build/*
 **/obj_dir/*
 /.vscode/
+/lib/
