+++ src/runtime/.gitignore
@@ -1,0 +1,1 @@
+Makefile_obj.mk
